[runtimes] Revert the libc++ __config_site change
This is a massive revert of the following commits (from most revent to oldest):

	2b9b7b5775a1d8fcd7aa5abaa8fc0bc303434f1a.
	821ca5a27dff7736fbe54b2174ba56421175df8c
	28270234f1478047e35879f4ba8838b47edfcc14
	9c5dfb2b74bc79fe2ace75a5b756d5ad0281a21c
	b5aa67446e01bd277727b05710a42e69ac41e74b
	680ba510201686802a23c48f183bb11c954cae82

After checking-in the __config_site change, a lot of things started breaking
due to widespread reliance on various aspects of libc++'s build, notably the
fact that we can include the headers from the source tree, but also reliance
on various "internal" CMake variables used by the runtimes build and compiler-rt.

These were unintended consequences of the change, and after two days, we
still haven't restored all the bots to being green. Instead, now that I
understand what specific areas this will blow up in, I should be able to
chop up the patch into smaller ones that are easier to digest.

See https://reviews.llvm.org/D89041 for more details on this adventure.